Enum ibr::ibd::redo::LogRecordTypes

#[repr(u8)]
pub enum LogRecordTypes {

    MLOG_SINGLE_REC_FLAG = 128,
    MLOG_1BYTE = 1,
    MLOG_2BYTES = 2,
    MLOG_4BYTES = 4,
    MLOG_8BYTES = 8,
    MLOG_REC_INSERT_8027 = 9,
    MLOG_REC_CLUST_DELETE_MARK_8027 = 10,
    MLOG_REC_SEC_DELETE_MARK = 11,
    MLOG_REC_UPDATE_IN_PLACE_8027 = 13,
    MLOG_REC_DELETE_8027 = 14,
    MLOG_LIST_END_DELETE_8027 = 15,
    MLOG_LIST_START_DELETE_8027 = 16,
    MLOG_LIST_END_COPY_CREATED_8027 = 17,
    MLOG_PAGE_REORGANIZE_8027 = 18,
    MLOG_PAGE_CREATE = 19,
    MLOG_UNDO_INSERT = 20,
    MLOG_UNDO_ERASE_END = 21,
    MLOG_UNDO_INIT = 22,
    MLOG_UNDO_HDR_REUSE = 24,
    MLOG_UNDO_HDR_CREATE = 25,
    MLOG_REC_MIN_MARK = 26,
    MLOG_IBUF_BITMAP_INIT = 27,
    MLOG_LSN = 28,
    MLOG_INIT_FILE_PAGE = 29,
    MLOG_WRITE_STRING = 30,
    MLOG_MULTI_REC_END = 31,
    MLOG_DUMMY_RECORD = 32,
    MLOG_FILE_CREATE = 33,
    MLOG_FILE_RENAME = 34,
    MLOG_FILE_DELETE = 35,
    MLOG_COMP_REC_MIN_MARK = 36,
    MLOG_COMP_PAGE_CREATE = 37,
    MLOG_COMP_REC_INSERT_8027 = 38,
    MLOG_COMP_REC_CLUST_DELETE_MARK_8027 = 39,
    MLOG_COMP_REC_SEC_DELETE_MARK = 40,
    MLOG_COMP_REC_UPDATE_IN_PLACE_8027 = 41,
    MLOG_COMP_REC_DELETE_8027 = 42,
    MLOG_COMP_LIST_END_DELETE_8027 = 43,
    MLOG_COMP_LIST_START_DELETE_8027 = 44,
    MLOG_COMP_LIST_END_COPY_CREATED_8027 = 45,
    MLOG_COMP_PAGE_REORGANIZE_8027 = 46,
    MLOG_ZIP_WRITE_NODE_PTR = 48,
    MLOG_ZIP_WRITE_BLOB_PTR = 49,
    MLOG_ZIP_WRITE_HEADER = 50,
    MLOG_ZIP_PAGE_COMPRESS = 51,
    MLOG_ZIP_PAGE_COMPRESS_NO_DATA_8027 = 52,
    MLOG_ZIP_PAGE_REORGANIZE_8027 = 53,
    MLOG_PAGE_CREATE_RTREE = 57,
    MLOG_COMP_PAGE_CREATE_RTREE = 58,
    MLOG_INIT_FILE_PAGE2 = 59,
    MLOG_TRUNCATE = 60,
    MLOG_INDEX_LOAD = 61,
    MLOG_TABLE_DYNAMIC_META = 62,
    MLOG_PAGE_CREATE_SDI = 63,
    MLOG_COMP_PAGE_CREATE_SDI = 64,
    MLOG_FILE_EXTEND = 65,
    MLOG_TEST = 66,
    MLOG_REC_INSERT = 67,
    MLOG_REC_CLUST_DELETE_MARK = 68,
    MLOG_REC_DELETE = 69,
    MLOG_REC_UPDATE_IN_PLACE = 70,
    MLOG_LIST_END_COPY_CREATED = 71,
    MLOG_PAGE_REORGANIZE = 72,
    MLOG_ZIP_PAGE_REORGANIZE = 73,
    MLOG_ZIP_PAGE_COMPRESS_NO_DATA = 74,
    MLOG_LIST_END_DELETE = 75,
    MLOG_LIST_START_DELETE = 76,
    UNDEF = 77,

}

types of a redo log record

Variants

MLOG_SINGLE_REC_FLAG = 128

if the mtr contains only one log record for one page, i.e., write_initial_log_record has been called only once, this flag is ORed to the type of that first log record

MLOG_1BYTE = 1

one byte is written

MLOG_2BYTES = 2

2 bytes …

MLOG_4BYTES = 4

4 bytes …

MLOG_8BYTES = 8

8 bytes …

MLOG_REC_INSERT_8027 = 9

Record insert

MLOG_REC_CLUST_DELETE_MARK_8027 = 10

Mark clustered index record deleted

MLOG_REC_SEC_DELETE_MARK = 11

Mark secondary index record deleted

MLOG_REC_UPDATE_IN_PLACE_8027 = 13

update of a record, preserves record field sizes

MLOG_REC_DELETE_8027 = 14

Delete a record from a page

MLOG_LIST_END_DELETE_8027 = 15

Delete record list end on index page

MLOG_LIST_START_DELETE_8027 = 16

Delete record list start on index page

MLOG_LIST_END_COPY_CREATED_8027 = 17

Copy record list end to a new created index page

MLOG_PAGE_REORGANIZE_8027 = 18

Reorganize an index page in ROW_FORMAT=REDUNDANT

MLOG_PAGE_CREATE = 19

Create an index page

MLOG_UNDO_INSERT = 20

Insert entry in an undo log

MLOG_UNDO_ERASE_END = 21

erase an undo log page end

MLOG_UNDO_INIT = 22

initialize a page in an undo log

MLOG_UNDO_HDR_REUSE = 24

reuse an insert undo log header

MLOG_UNDO_HDR_CREATE = 25

create an undo log header

MLOG_REC_MIN_MARK = 26

mark an index record as the predefined minimum record

MLOG_IBUF_BITMAP_INIT = 27

initialize an ibuf bitmap page

MLOG_LSN = 28

Current LSN

MLOG_INIT_FILE_PAGE = 29

this means that a file page is taken into use and the prior contents of the page should be ignored: in recovery we must not trust the lsn values stored to the file page. Note: it’s deprecated because it causes crash recovery problem in bulk create index, and actually we don’t need to reset page lsn in recv_recover_page_func() now.

MLOG_WRITE_STRING = 30

write a string to a page

MLOG_MULTI_REC_END = 31

If a single mtr writes several log records, this log record ends the sequence of these records

MLOG_DUMMY_RECORD = 32

dummy log record used to pad a log block full

MLOG_FILE_CREATE = 33

log record about creating an .ibd file, with format

MLOG_FILE_RENAME = 34

rename a tablespace file that starts with (space_id,page_no)

MLOG_FILE_DELETE = 35

delete a tablespace file that starts with (space_id,page_no)

MLOG_COMP_REC_MIN_MARK = 36

mark a compact index record as the predefined minimum record

MLOG_COMP_PAGE_CREATE = 37

create a compact index page

MLOG_COMP_REC_INSERT_8027 = 38

compact record insert

MLOG_COMP_REC_CLUST_DELETE_MARK_8027 = 39

mark compact clustered index record deleted

MLOG_COMP_REC_SEC_DELETE_MARK = 40

mark compact secondary index record deleted; this log record type is redundant, as MLOG_REC_SEC_DELETE_MARK is independent of the record format.

MLOG_COMP_REC_UPDATE_IN_PLACE_8027 = 41

update of a compact record, preserves record field sizes

MLOG_COMP_REC_DELETE_8027 = 42

delete a compact record from a page

MLOG_COMP_LIST_END_DELETE_8027 = 43

delete compact record list end on index page

MLOG_COMP_LIST_START_DELETE_8027 = 44

delete compact record list start on index page

MLOG_COMP_LIST_END_COPY_CREATED_8027 = 45

copy compact record list end to a new created index page

MLOG_COMP_PAGE_REORGANIZE_8027 = 46

reorganize an index page

MLOG_ZIP_WRITE_NODE_PTR = 48

write the node pointer of a record on a compressed non-leaf B-tree page

MLOG_ZIP_WRITE_BLOB_PTR = 49

write the BLOB pointer of an externally stored column on a compressed page

MLOG_ZIP_WRITE_HEADER = 50

write to compressed page header

MLOG_ZIP_PAGE_COMPRESS = 51

compress an index page

MLOG_ZIP_PAGE_COMPRESS_NO_DATA_8027 = 52

compress an index page without logging it’s image

MLOG_ZIP_PAGE_REORGANIZE_8027 = 53

reorganize a compressed page

MLOG_PAGE_CREATE_RTREE = 57

Create a R-Tree index page

MLOG_COMP_PAGE_CREATE_RTREE = 58

create a R-tree compact page

MLOG_INIT_FILE_PAGE2 = 59

this means that a file page is taken into use. We use it to replace MLOG_INIT_FILE_PAGE.

MLOG_TRUNCATE = 60

Table is being truncated. (Marked only for file-per-table), Disabled for WL6378

MLOG_INDEX_LOAD = 61

notify that an index tree is being loaded without writing redo log about individual pages

MLOG_TABLE_DYNAMIC_META = 62

log for some persistent dynamic metadata change

MLOG_PAGE_CREATE_SDI = 63

create a SDI index page

MLOG_COMP_PAGE_CREATE_SDI = 64

create a SDI compact page

MLOG_FILE_EXTEND = 65

Extend the space

MLOG_TEST = 66

Used in tests of redo log. It must never be used outside unit tests.

MLOG_REC_INSERT = 67

MLOG_REC_CLUST_DELETE_MARK = 68

MLOG_REC_DELETE = 69

MLOG_REC_UPDATE_IN_PLACE = 70

MLOG_LIST_END_COPY_CREATED = 71

MLOG_PAGE_REORGANIZE = 72

MLOG_ZIP_PAGE_REORGANIZE = 73

MLOG_ZIP_PAGE_COMPRESS_NO_DATA = 74

MLOG_LIST_END_DELETE = 75

MLOG_LIST_START_DELETE = 76

UNDEF = 77

biggest value (used in assertions)