Struct ibr::ibd::undo::UndoLogHeader

pub struct UndoLogHeader {

    pub addr: usize,
    pub buf: Arc<Bytes>,
    pub trx_id: u64,
    pub trx_no: u64,
    pub del_marks: u16,
    pub log_start: u16,
    pub undo_flags: Vec<UndoFlags>,
    pub undo_flags_bits: u8,
    pub dict_trans: u8,
    pub table_id: u64,
    pub next_log: u16,
    pub prev_log: u16,
    pub history_node: FlstNode,
    pub xa_trx_info: Option<XaTrxInfo>,

}

undo log header, see trx0undo.h

Fields
addr: usize

page address

buf: Arc<Bytes>

page data buffer

trx_id: u64

(8 bytes) TRX_UNDO_TRX_ID, Transaction id

trx_no: u64

(8 bytes) TRX_UNDO_TRX_NO, Transaction number of the transaction; defined only if the log is in a history list

del_marks: u16

(2 bytes) TRX_UNDO_DEL_MARKS, Defined only in an update undo log: true if the transaction may have done delete markings of records, and thus purge is necessary

log_start: u16

(2 bytes) TRX_UNDO_LOG_START, Offset of the first undo log record of this log on the header page; purge may remove undo log record from the log start, and therefore this is not necessarily the same as this log header end offset

undo_flags: Vec<UndoFlags>

see undo_flags_bits

undo_flags_bits: u8

(1 byte) TRX_UNDO_FLAGS, Transaction UNDO flags in one byte. This is backward compatible as earlier we were storing either 1 or 0 for TRX_UNDO_XID_EXISTS

dict_trans: u8

(1 byte), TRX_UNDO_DICT_TRANS, true if the transaction is a table create, index create, or drop transaction: in recovery the transaction cannot be rolled back in the usual way: a โrollbackโ rather means dropping the created or dropped table, if it still exists

table_id: u64

(8 bytes) TRX_UNDO_TABLE_ID, Id of the table if the preceding field is true. Note: deprecated

next_log: u16

(2 bytes) TRX_UNDO_NEXT_LOG, Offset of the next undo log header on this page, 0 if none

prev_log: u16

(2 bytes) TRX_UNDO_PREV_LOG, Offset of the previous undo log header on this page, 0 if none

history_node: FlstNode

(12 bytes) TRX_UNDO_HISTORY_NODE, If the log is put to the history list, the file list node is here

xa_trx_info: Option<XaTrxInfo>

(140 bytes) XA

Implementations

impl UndoLogHeader

pub fn new(addr: usize, buf: Arc<Bytes>) -> Self

const TRX_UNDO_FLAG_XID: u8 = 1u8

true if undo log header includes X/Open XA transaction identification XID

const TRX_UNDO_FLAG_GTID: u8 = 2u8

true if undo log header includes GTID information from replication

const TRX_UNDO_FLAG_XA_PREPARE_GTID: u8 = 4u8

true if undo log header includes GTID information for XA PREPARE

pub fn is_xid(&self) -> bool

pub fn is_gtid(&self) -> bool

pub fn is_xa_prepare_gtid(&self) -> bool