Module ibr::ibd::page

Structs

BasePage

Base Page Structure

C

Clean Bit

DoubleWriteBufferInfo

MySQL Double Write Buffer Info, see trx0sys.h, TRX_SYS_DOUBLEWRITE

F

Free Bit

FSegHeader

File Segment Header, see fsp0types.h/page0types.h

FilAddr

File Address

FilePageHeader

FIL Header, see fil0types.h

FilePageTrailer

FIL Trailer, see fil0types.h

FileSpaceFlags

FSP Header, see fsp0types.h, FSP_FLAGS_WIDTH_xxxx

FileSpaceHeader

FSP Header, see fsp0fsp.h

FileSpaceHeaderPageBody

File Space Header Page

FlstBaseNode

File List Base Node

FlstNode

File List Node

INodeEntry

INode Entry, see fsp0fsp.h

INodePageBody

File Segment Inode, see fsp0fsp.h

IndexHeader

Index Page Header, see page0types.h

IndexPageBody

Index Page

LogInfo

MySQL Log Info, see trx0sys.h

RSegArrayPageBody

Rollback Segment Header Page, see trx0rseg.h

RSegHeaderPageBody

Rollback Segment Header Page, see trx0rseg.h

RSegInfo

Rollback segment information, see trx0sys.h

RollbackSegmentHeader

Rollback Segment Header, see trx0rseg.h

SdiMetaData

SDI Meta Data

SdiPageBody

SDI Index Page, see ibd2sdi.cc

TrxSysPageBody

Transaction System Page, see trx0sys.h

UndoLogPageBody

UNDO Log Page, see trx0undo.h

UndoPageHeader

Undo Page Header, see trx0undo.h

UndoSegmentHeader

Undo Segment Header, see trx0undo.h

XDesEntry

Extent Descriptor Entry, see fsp0fsp.h

XDesPageBody

Extent Descriptor Page, much-like FileSpaceHeaderPageBody, except zero-fill the fsp_hdr

Enums

PageDirections

Page Direction

PageFormats

Page Format, redundant or compact

PageNumber

Page Number

PageTypes

MySQL Page Type, see fil0fil.h

SpaceId

Tablespace ID

UndoPageStates

States of an undo log segment

UndoPageTypes

Types of an undo log segment

XDesStates

Extent Descriptor State

Constants

DICT_SPACE_ID

EXTENT_PAGE_NUM

FIL_HEADER_SIZE

FIL_TRAILER_SIZE

FRAG_ARR_ENTRY_SIZE

FSEG_FRAG_ARR_OFFSET

FSEG_MAGIC_NUMBER

FSP_DICT_HDR_PAGE_NO

data dictionary header page, in tablespace 0

FSP_FIRST_RSEG_PAGE_NO

first rollback segment page, in tablespace 0

FSP_HEADER_SIZE

FSP_IBUF_HEADER_PAGE_NO

The following pages exist in the system tablespace (space 0). insert buffer header page, in tablespace 0

FSP_IBUF_TREE_ROOT_PAGE_NO

insert buffer B-tree root page in tablespace 0; The ibuf tree root page number in tablespace 0; its fseg inode is on the page number FSP_FIRST_INODE_PAGE_NO

FSP_TRX_SYS_PAGE_NO

transaction system header, in tablespace 0

INDEX_FSEG_HEADER_SIZE

INDEX_HEADER_SIZE

INF_PAGE_BYTE_OFF

INNODB_TEMP_SPACE_ID

INODE_ENTRY_ARR_COUNT

INODE_ENTRY_MAX_COUNT

INODE_ENTRY_SIZE

INODE_FLST_NODE_SIZE

INVALID_SPACE_ID

PAGE_DIR_ENTRY_SIZE

PAGE_NONE

RECORD_HEADER_SIZE

REDO_LOG_SPACE_ID

SUP_PAGE_BYTE_OFF

TEMP_SPACE_ID_MAX

TEMP_SPACE_ID_MIN

TRX_RSEG_N_SLOTS

TRX_RSEG_SLOT_SIZE

TRX_SYS_DOUBLEWRITE

The offset of the doublewrite buffer header on the trx system header page

TRX_SYS_DOUBLEWRITE_MAGIC_N

TRX_SYS_DOUBLEWRITE_SPACE_ID_STORED_N

TRX_SYS_MYSQL_LOG_INFO

TRX_SYS_MYSQL_LOG_NAME_LEN

TRX_SYS_N_RSEGS

TRX_SYS_SPACE_ID

UNDO_SPACE_ID_MAX

UNDO_SPACE_ID_MIN

UNIV_PAGE_SIZE

universal page size, 16k

XDES_ENTRY_MAX_COUNT

XDES_ENTRY_SIZE

XDES_PAGE_COUNT

Traits

BasePageBody

Base Page Body