Module ibr::ibd::redo

Structs

LogBlock

log block, see log0constants.h

LogCheckpoint

log checkpoint, see log0constants.h

LogFile

log file, see log0constants.h

LogFileHeader

log file header, see log0constants.h

LogRecord

log record, see recv_parse_log_rec(…)

LogRecordHeader

log record header, see mtr0log.ic, mlog_write_initial_log_record_low(…)

RedoLogIndexInfo

redo log index info, see mlog_parse_index(…)

RedoRecForFileDelete

log record payload for file delete, see fil_tablespace_redo_delete(…)

RedoRecForNByte

log record payload for nByte, see mlog_parse_nbytes(…)

RedoRecForRecordClusterDeleteMark

log record payload for marking a clustered index record deleted, see btr_cur_parse_del_mark_set_clust_rec(…)

RedoRecForRecordDelete

log record payload for delete record, see page_cur_parse_delete_rec(…)

RedoRecForRecordInsert

log record payload for insert record, see page_cur_parse_insert_rec(…)

RedoRecForRecordSecIndexDeleteMark

log record payload for marking a secondary index record deleted, see btr_cur_parse_del_mark_set_sec_rec(…)

RedoRecForRecordUpdateInPlace

log record payload for update record in-place, see btr_cur_parse_update_in_place(…)

RedoRecForTableDynamicMeta

log record payload for log for some persistent dynamic metadata change, see

RedoRecForUndoInsert

log record payload for parsing a redo log record of adding an undo log record, see trx_undo_parse_add_undo_rec(…)

RedoRecForUndoPageHeader

log record payload for parsing the redo log entry of an undo log page header create or reuse, see trx_undo_parse_page_header(…)

RedoRecUpdatedField

redo record updated fields

Enums

Blocks

BtrModeFlags

IndexFieldFixed

index field fixed

IndexFieldNullable

index field nullable

IndexInfoFlags

index flags

LogRecordTypes

types of a redo log record

PersistentTypes

Persistent dynamic metadata type, there should be 1 to 1 relationship between the metadata and the type. Please keep them in order so that we can iterate over it

RedoRecordPayloads

redo record payload, see recv_parse_or_apply_log_rec_body(…)

Constants

BTR_CREATE_FLAG 🔒

BTR_KEEP_IBUF_BITMAP 🔒

BTR_KEEP_POS_FLAG 🔒

BTR_KEEP_SYS_FLAG 🔒

BTR_NO_LOCKING_FLAG 🔒

BTR_NO_UNDO_LOG_FLAG 🔒

LOG_BLOCK_MAX_NO

Maximum allowed block’s number (stored in hdr_no) increased by 1.

OS_FILE_LOG_BLOCK_SIZE

The next value should be smaller or equal to the smallest sector size used on any disk. A log block is required to be a portion of disk which is written so that if the start and the end of a block get written to disk, then the whole block gets written. This should be true even in most cases of a crash: if this fails for a log block, then it is equivalent to a media failure in the log.

Functions

log_block_convert_lsn_to_epoch_no

log_block_convert_lsn_to_hdr_no

log_block_guess_lsn

parse_mode_flags

parse mode flags